+++ docs/changelog.rst
@@ -5,6 +5,7 @@ Changelog
 * :bug:`-` Transaction decoding will no longer fail when encountering a certain rare case of problematic spam tokens.
 * :bug:`-` Some specific 0x settler swaps in Optimism will now be properly decoded.
 * :bug:`-` When exporting history as CSV the events will now properly appear sorted by timestamp.
+* :bug:`10087` Users will now be able to use the electron app wallet bridge when using Windows.
 
 * :release:`1.39.0 <2025-06-04>`
 * :feature:`-` DigixDAO DGD refunds will now be properly decoded.